Move main-process files under /main, and /renderer respectively
