+++ package.json
@@ -3,7 +3,7 @@
   "productName": "Helmet 4.0 UI",
   "version": "1.1.0-SNAPSHOT",
   "description": "Helmet 4.0 Model System UI",
-  "main": "src/index.js",
+  "main": "src/main/index.js",
   "scripts": {
     "start": "electron-forge start",
     "package": "electron-forge package",
